Fix config file check test
The test was not actually testing anything, as shown my the coverage
report. Tweak the main function to return instead of exit the program,
and change the test to test the main function instead of the class,
since that is where we are testing for the config file not being found.

Add testing against Python 3.6.

Create a requirements.txt.